+++ applications/fishing-map/src/features/datasets/datasets.utils.ts
@@ -5,11 +5,7 @@ import { capitalize, sortFields } from 'utils/shared'
 import { t } from 'features/i18n/i18n'
 import { PUBLIC_SUFIX, FULL_SUFIX, PRIVATE_SUFIX } from 'data/config'
 import { getDatasetNameTranslated } from 'features/i18n/utils'
-import {
-  DEFAULT_FISHING_DATAVIEW_ID,
-  DEFAULT_PRESENCE_DATAVIEW_ID,
-  DEFAULT_VIIRS_DATAVIEW_ID,
-} from 'data/workspaces'
+import { FISHING_DATAVIEW_ID, PRESENCE_DATAVIEW_ID, VIIRS_DATAVIEW_ID } from 'data/workspaces'
 
 export type SupportedDatasetSchema =
   | 'flag'
@@ -48,11 +44,11 @@ export const getDatasetTitleByDataview = (
     ? dataview.datasets?.filter((d) => dataview.config?.datasets?.includes(d.id))
     : dataview.datasets
   let datasetTitle = dataview.name || ''
-  if (dataviewInstance.dataviewId === DEFAULT_FISHING_DATAVIEW_ID) {
+  if (dataviewInstance.dataviewId === FISHING_DATAVIEW_ID) {
     datasetTitle = t(`common.apparentFishing`, 'Apparent Fishing Effort')
-  } else if (dataviewInstance.dataviewId === DEFAULT_PRESENCE_DATAVIEW_ID) {
+  } else if (dataviewInstance.dataviewId === PRESENCE_DATAVIEW_ID) {
     datasetTitle = t(`common.presence`, 'Vessel presence')
-  } else if (dataviewInstance.dataviewId === DEFAULT_VIIRS_DATAVIEW_ID) {
+  } else if (dataviewInstance.dataviewId === VIIRS_DATAVIEW_ID) {
     datasetTitle = t(`common.viirs`, 'Night light detections (VIIRS)')
   } else if (activeDatasets) {
     if (hasDatasetsConfig && activeDatasets?.length !== 1) {